requirements.txt
yt-dlp
telethon
python-dotenv
